requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.3
django-filter==2.2.0
django-stubs==1.4.0
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
djangorestframework-stubs==1.1.0
entrypoints==0.3
flake8==3.7.9
importlib-metadata==1.5.0
Markdown==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
seval==1.1.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.8
whitenoise==5.0.1
zipp==3.0.0
